backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
kerykeion>=1.1.28
httpx>=0.27.0
pydantic[email]>=2.7.0
python-multipart>=0.0.9
google-generativeai>=0.8.0
groq>=0.5.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
skyfield>=1.49
pytz
reportlab
numpy
python-jose[cryptography]
bcrypt
matplotlib
asyncpg
gunicorn
psycopg2-binary
pyswisseph
